requirements.txt
Flask==2.1.1
Flask-Session==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.26.1
cryptography==36.0.1
